Add error messages in each API

+++ src/routes/deleteUser.js
@@ -2,13 +2,25 @@ const { User } = require("@database/sequelize");
 
 module.exports = (app) => {
   app.delete("/api/users/:id", (req, res) => {
-    User.findByPk(req.params.id).then((user) => {
-      const deletedUser = user;
+    User.findByPk(req.params.id)
+      .then((user) => {
+        if (user === null) {
+          const message =
+            "The requested user does not exist. Please try again with a different identifier.";
+          return res.status(404).json({ message: message });
+        }
 
-      return User.destroy({ where: { id: user.id } }).then(() => {
-        const message = `The user (id:${deletedUser.id}) has been deleted.`;
-        res.json({ message, data: deletedUser });
+        const deletedUser = user;
+
+        return User.destroy({ where: { id: user.id } }).then(() => {
+          const message = `The user (id:${deletedUser.id}) has been deleted.`;
+          res.json({ message, data: deletedUser });
+        });
+      })
+      .catch((error) => {
+        const message =
+          "The user could not be updated. Please try again in a few moments.";
+        res.status(500).json({ message, error: error.message });
       });
-    });
   });
 };
